+++ src/types/@microsoft-botframework-webchat-adapter-azure-communication-chat.d.ts
@@ -1,0 +1,5 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+// typings/@microsoft botframework-webchat-adapter-azure-communication-chat.d.ts
+declare module '@microsoft/botframework-webchat-adapter-azure-communication-chat' {
+    export function createACSAdapter(...args: any[]): any;
+}
