--- src/types/acs_webchat-chat-adapter.d.ts
@@ -1,5 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-// typings/acs_webchat-chat-adapter.d.ts
-declare module 'acs_webchat-chat-adapter' {
-    export function createACSAdapter(...args: any[]): any;
-}
