--- src/utils/__mocks__/acs_webchat-chat-adapter.js
@@ -1,19 +1,0 @@
-
-const acs_webchat_chat_adapter = jest.createMockFromModule('acs_webchat-chat-adapter');
-
-acs_webchat_chat_adapter.createACSAdapter = () => {
-    return {
-    end : () => {
-        console.log("end");
-    },
-    fileManager : {
-        fileScanner : {
-            end : () => {
-                console.log("file scanner end");
-            },
-        },
-    }
-};
-
-};
-module.exports = acs_webchat_chat_adapter;
